Stopped duplicates from being added into the db
Removed duplicate insertion from the database in two ways.
(1) When the list of users is created from the comments, only
insert new users into the list. This is because someone can reply or
double comment

(2) Early on, get a list of the unique id's from the database. Then
after validating (1) perform a regex and extract the user_id from the
user's url for each unique user. Cross check this against the unique_ids
that have already been extracted from the database and only queue up
new ids to be inserted into the database.

If I give the program on b02 a repeat article with 15 comments withouth having
to load a new page it takes 18s to validate that it has seen all users and
cloes the db and does not insert anything new.