Fixed multi-user per article bug
Correctly implemented the removal of multiple users per article.
The problem was looking for unique user from the selnium comment
block rather than the href that had the actual user URLs in them.

Fixed and made sure that the check for duplicate users in the python
list is done first before the database check.