Redo-ing  previous buggy duplicate ID fix
Previous version did it incorrectly beceause I mis-read wwhat
was happening when the regex extracts data. It is the get_atttribute('href')
that is important no the result of regex.

Checked multiple times, this correctly ignores duplicates within the
same article. Went from 6000->900 (comments->unique) in one article
and 1600->400 in another.